♻️ refactor(core): 合并执行路径并精简 core 架构
