--- qqmusic_api/core/executors/__init__.py
@@ -1,9 +1,0 @@
-"""请求执行器包."""
-
-from .cgi import CgiExecutor
-from .http import HttpExecutor
-
-__all__ = [
-    "CgiExecutor",
-    "HttpExecutor",
-]
