+++ qqmusic_api/core/transport.py
@@ -4,12 +4,16 @@
 响应体均已就绪; 流式请求 (kwargs 携带 stream=True) 返回到响应头就绪,
 响应体的延迟读取与关闭由调用者负责. 收到响应的调用者必须 ``release``
 或明确移交所有权.
+
+批量模型: ``request_many`` 按 **物理请求** 归属结果与异常 — 返回与输入
+顺序一致的结果列表, 元素为 ``RawResponse`` 或 ``TransportError``, 不因
+单个物理请求失败而丢弃其他请求的结果.
 """
 
 import contextlib
 from collections.abc import Mapping, Sequence
 from dataclasses import dataclass, field
-from typing import Any, Protocol, cast, runtime_checkable
+from typing import Any, Protocol, TypeAlias, cast, runtime_checkable
 
 import anyio
 from niquests import AsyncSession, AsyncTokenBucketLimiter, RetryConfiguration
@@ -17,8 +21,6 @@ from niquests import PreparedRequest as NiquestsPreparedRequest
 from niquests.exceptions import RequestException, Timeout
 from niquests.models import Response
 from niquests.typing import AsyncHookType, ProxyType, TLSClientCertType, TLSVerifyType
-
-from .runtime import DEFAULT_MAX_CONCURRENCY
 
 __all__ = [
     "HttpRawResponse",
@@ -29,10 +31,16 @@ __all__ = [
     "Transport",
     "TransportError",
     "TransportTimeout",
+    "send_many",
 ]
 
 HttpRawResponse = Response
 """底层 HTTP 响应的公开类型别名. 供请求描述符在 ``disable_parse`` 场景标注结果类型."""
+
+DEFAULT_MAX_CONCURRENCY = 20
+
+BatchOutcome: TypeAlias = "RawResponse | TransportError"
+"""单个物理请求的批量结果: 响应或归属到该请求的传输异常."""
 
 
 class TransportError(Exception):
@@ -113,9 +121,49 @@ class Transport(Protocol):
 class MultiplexTransport(Protocol):
     """支持先提交多个请求、再集中解析响应的传输扩展."""
 
-    async def request_many(self, requests: Sequence[PreparedRequest]) -> list[RawResponse]:
-        """批量提交请求并在全部响应就绪后按输入顺序返回."""
+    async def request_many(self, requests: Sequence[PreparedRequest]) -> "list[BatchOutcome]":
+        """批量提交请求并按输入顺序返回逐请求结果 (响应或异常)."""
         ...
+
+
+class _CapacityLimiter:
+    """支持原子多许可获取的容量限制器.
+
+    ``acquire(n)`` 等待到空闲容量不少于 n 后一次性扣除, 等待期间
+    不持有任何许可 — 多个批量请求不会互相持有部分许可而死锁,
+    小请求也可以在批量等待期间利用零散空闲容量.
+    """
+
+    def __init__(self, total: int) -> None:
+        """初始化容量限制器.
+
+        Args:
+            total: 总容量.
+        """
+        self._total = total
+        self._used = 0
+        self._condition = anyio.Condition()
+
+    async def acquire(self, amount: int = 1) -> None:
+        """获取指定数量的许可, 必要时等待.
+
+        Args:
+            amount: 需要的许可数量.
+        """
+        async with self._condition:
+            while self._used + amount > self._total:
+                await self._condition.wait()
+            self._used += amount
+
+    async def release(self, amount: int = 1) -> None:
+        """归还指定数量的许可并唤醒等待者.
+
+        Args:
+            amount: 归还的许可数量.
+        """
+        async with self._condition:
+            self._used -= amount
+            self._condition.notify_all()
 
 
 def _map_transport_exception(exc: RequestException) -> TransportError:
@@ -146,12 +194,78 @@ async def _release_raw(response: RawResponse) -> None:
         await result
 
 
+async def send_many(
+    transport: Transport,
+    requests: "Sequence[PreparedRequest]",
+    *,
+    max_concurrency: int,
+) -> "list[BatchOutcome]":
+    """批量发送辅助函数. 两个执行器共用的唯一批量入口.
+
+    支持批量能力 (``MultiplexTransport``) 的传输直接委托
+    ``request_many``; 其余实现以有限并发 worker 回退, 回退逻辑
+    全库仅此一份. 结果按物理请求归属, 异常不跨请求扩散.
+
+    Args:
+        transport: 传输边界.
+        requests: 待发送的传输请求序列.
+        max_concurrency: 回退路径的并发 worker 上限.
+
+    Returns:
+        与输入顺序一致的逐请求结果列表.
+    """
+    if isinstance(transport, MultiplexTransport):
+        return await transport.request_many(requests)
+    return await _send_many_fallback(transport, requests, max_concurrency)
+
+
+async def _send_many_fallback(
+    transport: Transport,
+    requests: "Sequence[PreparedRequest]",
+    max_concurrency: int,
+) -> "list[BatchOutcome]":
+    """无批量能力传输的有限并发回退.
+
+    Args:
+        transport: 传输边界.
+        requests: 待发送的传输请求序列.
+        max_concurrency: 并发 worker 上限.
+
+    Returns:
+        与输入顺序一致的逐请求结果列表.
+    """
+    outcomes: list[BatchOutcome] = [TransportError("未发送")] * len(requests)
+    pending = iter(list(enumerate(requests)))
+    pending_lock = anyio.Lock()
+
+    async def _worker() -> None:
+        while True:
+            async with pending_lock:
+                entry = next(pending, None)
+            if entry is None:
+                return
+            position, request = entry
+            try:
+                outcomes[position] = await transport.request(request)
+            except TransportError as exc:
+                outcomes[position] = exc
+            except Exception as exc:
+                outcomes[position] = TransportError(str(exc))
+
+    async with anyio.create_task_group() as task_group:
+        for _ in range(min(max_concurrency, len(requests)) or 1):
+            task_group.start_soon(_worker)
+
+    return outcomes
+
+
 class NiquestsTransport:
-    """基于 niquests AsyncSession 的单物理请求传输实现.
+    """基于 niquests AsyncSession 的传输实现.
 
     拥有底层会话与代理, 证书, hooks, verify 等发送配置; 配置在每次
-    ``request`` 进入时读取. 共享容量信号量覆盖从获取连接到响应就绪的
-    全程 (缓冲) 或到响应头就绪 (流式); 交付后的流不计入在途请求数.
+    请求进入时读取. 保留 niquests 多路复用工作流: 先提交一批 lazy
+    请求, 再集中 gather. 容量信号量在提交前整块获取 — 不存在多个
+    批次各持有部分许可互相等待的死锁.
 
     适配器隔离能力已验证 (Task A): 单请求取消只影响自身, 客户端取消时
     关闭底层连接, 同会话其余请求与后续复用不受影响.
@@ -202,15 +316,11 @@ class NiquestsTransport:
         self.cert = cert
         self.verify = verify
         self.hooks = hooks
-        self._max_concurrency = max_concurrency
-        self._capacity = anyio.Semaphore(max_concurrency)
+        self._capacity = _CapacityLimiter(max_concurrency)
         self._closed = False
 
     async def request(self, request: PreparedRequest) -> RawResponse:
         """执行单个物理请求并返回响应.
-
-        缓冲请求返回时状态与响应体已就绪 (niquests 在发送阶段即等待完整
-        响应); 流式请求返回到响应头就绪.
 
         Args:
             request: 准备完成的传输请求.
@@ -222,35 +332,60 @@ class NiquestsTransport:
             TransportTimeout: 请求超时.
             TransportError: 其他网络异常.
         """
-        responses = await self.request_many([request])
-        return responses[0]
+        outcome = (await self.request_many([request]))[0]
+        if isinstance(outcome, TransportError):
+            raise outcome
+        return outcome
 
-    async def request_many(self, requests: Sequence[PreparedRequest]) -> list[RawResponse]:
+    async def request_many(self, requests: Sequence[PreparedRequest]) -> "list[BatchOutcome]":
         """分块提交 lazy 请求并集中解析响应.
 
-        每个分块先完成全部 ``AsyncSession.request`` 调用, 再执行一次
-        ``AsyncSession.gather(*responses)``, 保留 niquests 的多路复用工作流.
+        每个分块在提交前 **整块获取** 容量许可 (避免与其他批次互相
+        等待对方持有的部分许可), 随后完成全部 ``AsyncSession.request``
+        提交, 再执行一次 ``AsyncSession.gather``. 单个物理请求失败只
+        归属到该请求; 集中解析失败归属到其中仍未就绪的请求. 取消时
+        尽力释放已收集的全部响应.
 
         Args:
             requests: 待提交的传输请求序列.
 
         Returns:
-            与输入顺序一致的已就绪响应列表.
-
-        Raises:
-            TransportTimeout: 请求或集中解析超时.
-            TransportError: 其他网络异常.
+            与输入顺序一致的逐请求结果列表.
         """
-        results: list[RawResponse] = []
         items = list(requests)
-        for start in range(0, len(items), self._max_concurrency):
-            chunk = items[start : start + self._max_concurrency]
-            responses: list[RawResponse] = []
-            try:
-                for request in chunk:
-                    await self._capacity.acquire()
-                    try:
-                        response = await self._client.request(
+        outcomes: list[BatchOutcome] = [TransportError("未发送")] * len(items)
+        collected: list[RawResponse] = []
+        try:
+            for start in range(0, len(items), DEFAULT_MAX_CONCURRENCY):
+                chunk = items[start : start + DEFAULT_MAX_CONCURRENCY]
+                chunk_outcomes = await self._submit_chunk(chunk)
+                outcomes[start : start + len(chunk)] = chunk_outcomes
+                collected.extend(response for response in chunk_outcomes if isinstance(response, Response))
+        except BaseException:
+            # 后续分块失败时, 释放之前分块已收集但尚未交付的响应.
+            await self._discard(collected)
+            raise
+        return outcomes
+
+    async def _submit_chunk(self, chunk: Sequence[PreparedRequest]) -> "list[BatchOutcome]":
+        """提交单个容量分块并集中解析.
+
+        Args:
+            chunk: 本分块的传输请求.
+
+        Returns:
+            与分块顺序一致的逐请求结果列表.
+        """
+        chunk_outcomes: list[BatchOutcome] = [TransportError("未发送")] * len(chunk)
+        submitted: list[tuple[int, Response]] = []
+        await self._capacity.acquire(len(chunk))
+
+        try:
+            for position, request in enumerate(chunk):
+                try:
+                    response = cast(
+                        "Response",
+                        await self._client.request(
                             request.method,
                             request.url,
                             **dict(request.kwargs),
@@ -258,32 +393,56 @@ class NiquestsTransport:
                             hooks=self.hooks,
                             cert=self.cert,
                             verify=self.verify,
-                        )
-                    except BaseException:
-                        self._capacity.release()
-                        raise
-                    responses.append(response)
+                        ),
+                    )
+                    submitted.append((position, response))
+                    chunk_outcomes[position] = response
+                except (Timeout, RequestException) as exc:  # noqa: PERF203
+                    chunk_outcomes[position] = _map_transport_exception(exc)
+                except Exception as exc:
+                    # 取消等 BaseException 不包装, 直接传播.
+                    chunk_outcomes[position] = TransportError(str(exc))
 
-                lazy_responses = [response for response in responses if getattr(response, "lazy", False)]
-                if lazy_responses:
-                    await self._client.gather(*cast("list[Response]", lazy_responses))
-            except Timeout as exc:
-                await self._discard_unresolved(responses)
-                raise TransportTimeout(str(exc)) from exc
-            except RequestException as exc:
-                await self._discard_unresolved(responses)
-                raise TransportError(str(exc)) from exc
-            except BaseException:
-                await self._discard_unresolved(responses)
-                raise
-            finally:
-                for _ in responses:
-                    self._capacity.release()
-            results.extend(responses)
-        return results
+            lazy_pairs = [(position, response) for position, response in submitted if getattr(response, "lazy", False)]
+            if lazy_pairs:
+                try:
+                    await self._client.gather(*[response for _, response in lazy_pairs])
+                except (Timeout, RequestException) as exc:
+                    await self._fail_unresolved(chunk_outcomes, lazy_pairs, _map_transport_exception(exc))
+                except Exception as exc:
+                    # 取消等 BaseException 不包装, 直接传播.
+                    await self._fail_unresolved(chunk_outcomes, lazy_pairs, TransportError(str(exc)))
+        except BaseException:
+            # 外层取消等异常: 尽力释放本分块已收集的响应后继续传播.
+            await self._discard([response for _, response in submitted])
+            raise
+        finally:
+            await self._capacity.release(len(chunk))
 
-    async def _discard_unresolved(self, responses: Sequence[RawResponse]) -> None:
-        """尽力关闭未能集中解析完成的响应."""
+        return chunk_outcomes
+
+    @staticmethod
+    async def _fail_unresolved(
+        chunk_outcomes: "list[BatchOutcome]",
+        lazy_pairs: "list[tuple[int, Response]]",
+        error: TransportError,
+    ) -> None:
+        """将集中解析失败归属到仍未就绪的响应, 并尽力释放它们.
+
+        Args:
+            chunk_outcomes: 本分块的结果列表 (原地更新).
+            lazy_pairs: 集中解析前仍处于 lazy 状态的 (位置, 响应) 对.
+            error: 集中解析抛出的传输异常.
+        """
+        for position, _response in lazy_pairs:
+            chunk_outcomes[position] = error
+        with anyio.CancelScope(shield=True):
+            for _position, response in lazy_pairs:
+                with contextlib.suppress(Exception):
+                    await _release_raw(response)
+
+    async def _discard(self, responses: Sequence[RawResponse]) -> None:
+        """尽力释放一批响应 (屏蔽取消)."""
         with anyio.CancelScope(shield=True):
             for response in responses:
                 with contextlib.suppress(Exception):
